javais fais de la merde a gogo hihi

+++ srcs/classes/Response.cpp
@@ -91,7 +91,7 @@ std::string	Response::Header(void)
 		header << "Content-Length: " << this->body.size() << "\r\n";
 	}
 	header << "Connection: " << this->GetConnection() << "\r\n";
-	for (int i = 0; i < headers.size(); i++) {
+	for (size_t i = 0; i < headers.size(); i++) {
 		header << headers[i] << "\r\n";
 	}
 	header << "\r\n";
